вынесла функции в отдельный файл
